wip: refactoring rup
